move to profile.go
