+++ pkg/collector/corechecks/snmp/internal/checkconfig/config.go
@@ -538,35 +538,6 @@ func NewCheckConfig(rawInstance integration.Data, rawInitConfig integration.Data
 	return c, nil
 }
 
-func getProfiles(initConfigProfiles profile.ProfileConfigMap) (profile.ProfileConfigMap, error) {
-	var profiles profile.ProfileConfigMap
-	if len(initConfigProfiles) > 0 {
-		// TODO: [PERFORMANCE] Load init config custom profiles once for all integrations
-		//   There are possibly multiple init configs
-		customProfiles, err := loadProfiles(initConfigProfiles)
-		if err != nil {
-			return nil, fmt.Errorf("failed to load initConfig profiles: %s", err)
-		}
-		profiles = customProfiles
-	} else if profileBundleFileExist() {
-		defaultProfiles, err := loadBundleJSONProfiles()
-		if err != nil {
-			return nil, fmt.Errorf("failed to load bundle json profiles: %s", err)
-		}
-		profiles = defaultProfiles
-	} else {
-		defaultProfiles, err := loadYamlProfiles()
-		if err != nil {
-			return nil, fmt.Errorf("failed to load yaml profiles: %s", err)
-		}
-		profiles = defaultProfiles
-	}
-	for _, profileDef := range profiles {
-		profiledefinition.NormalizeMetrics(profileDef.Definition.Metrics)
-	}
-	return profiles, nil
-}
-
 func (c *CheckConfig) getResolvedSubnetName() string {
 	var resolvedSubnet string
 	if c.Network != "" {
